requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
pyspark==3.5.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
click==8.1.7
loguru==0.7.2
prometheus-client==0.19.0 